Add test & rename example package
